requirements.txt
numpy<2.0.0
pandas>=2.0.0
scikit-surprise>=1.1.3
scikit-learn>=1.2.0
joblib>=1.2.0
torch>=2.0.0
transformers>=4.30.0